move a few more components, fix tests
