+++ src/cdk/index.ts
@@ -1,2 +1,3 @@
 export * from './construct-lib';
+export * from './jsii-docgen';
 export * from './jsii-project';
