+++ src/javascript/index.ts
@@ -1,6 +1,7 @@
 export * from './bundler';
 export * from './eslint';
 export * from './jest';
+export * from './node-package';
 export * from './node-project';
 export * from './npm-config';
 export * from './upgrade-dependencies';
