+++ src/javascript/node-package.ts
@@ -2,14 +2,14 @@ import { join } from 'path';
 import { parse as urlparse } from 'url';
 import { accessSync, constants, existsSync, readdirSync, readJsonSync } from 'fs-extra';
 import * as semver from 'semver';
-import { resolve as resolveJson } from './_resolve';
-import { Component } from './component';
-import { DependencyType } from './dependencies';
-import { JsonFile } from './json';
-import { Project } from './project';
-import { isAwsCodeArtifactRegistry } from './release/publisher';
-import { Task } from './task';
-import { exec, isTruthy, sorted, writeFile } from './util';
+import { resolve as resolveJson } from '../_resolve';
+import { Component } from '../component';
+import { DependencyType } from '../dependencies';
+import { JsonFile } from '../json';
+import { Project } from '../project';
+import { isAwsCodeArtifactRegistry } from '../release';
+import { Task } from '../task';
+import { exec, isTruthy, sorted, writeFile } from '../util';
 
 const UNLICENSED = 'UNLICENSED';
 const DEFAULT_NPM_REGISTRY_URL = 'https://registry.npmjs.org/';
