+++ src/typescript/typescript.ts
@@ -1,14 +1,11 @@
 import * as path from 'path';
 import { PROJEN_DIR, PROJEN_RC } from '../common';
 import { Component } from '../component';
-import { NodeProject, NodeProjectOptions } from '../javascript';
-import { Eslint, EslintOptions } from '../javascript/eslint';
-import { TypeScriptCompilerOptions, TypescriptConfig, TypescriptConfigOptions } from '../javascript/typescript-config';
+import { Eslint, EslintOptions, NodeProject, NodeProjectOptions, TypeScriptCompilerOptions, TypescriptConfig, TypescriptConfigOptions } from '../javascript';
 import { SampleDir } from '../sample-file';
 import { Task } from '../task';
 import { TextFile } from '../textfile';
-import { Projenrc as ProjenrcTs, ProjenrcOptions as ProjenrcTsOptions } from '../typescript/projenrc';
-import { TypedocDocgen } from '../typescript/typescript-typedoc';
+import { Projenrc as ProjenrcTs, ProjenrcOptions as ProjenrcTsOptions, TypedocDocgen } from '../typescript';
 
 export interface TypeScriptProjectOptions extends NodeProjectOptions {
   /**
