+++ test/awscdk-construct.test.ts
@@ -2,7 +2,7 @@ import { mkdirSync, writeFileSync } from 'fs';
 import { join } from 'path';
 import { awscdk } from '../src';
 import { AwsCdkConstructLibrary, AwsCdkConstructLibraryOptions } from '../src/awscdk';
-import { NpmAccess } from '../src/node-package';
+import { NpmAccess } from '../src/javascript';
 import { mkdtemp, synthSnapshot } from '../src/util/synth';
 
 describe('constructs dependency selection', () => {
