+++ test/github/auto-approve.test.ts
@@ -1,5 +1,5 @@
 import { AutoApprove } from '../../src/github/auto-approve';
-import { NodeProject, NodeProjectOptions } from '../../src/node-project';
+import { NodeProject, NodeProjectOptions } from '../../src/javascript';
 import { synthSnapshot } from '../../src/util/synth';
 
 describe('auto-approve', () => {
