+++ test/github/dependabot.test.ts
@@ -1,5 +1,5 @@
 import { Dependabot, DependabotRegistryType } from '../../src/github';
-import { NodeProject, NodeProjectOptions } from '../../src/node-project';
+import { NodeProject, NodeProjectOptions } from '../../src/javascript';
 import { synthSnapshot } from '../../src/util/synth';
 
 describe('dependabot', () => {
