+++ test/github/task-workflow.test.ts
@@ -1,5 +1,5 @@
 import { TaskWorkflow } from '../../src/github/task-workflow';
-import { Task } from '../../src/tasks';
+import { Task } from '../../src/task';
 import { synthSnapshot, TestProject } from '../../src/util/synth';
 
 describe('task-workflow', () => {
