+++ test/json/projenrc.test.ts
@@ -1,5 +1,5 @@
 import { renderProjenNewOptions } from '../../src/javascript/render-options';
-import { Projenrc } from '../../src/json/index';
+import { Projenrc } from '../../src/projenrc-json';
 import { synthSnapshot, TestProject } from '../../src/util/synth';
 
 test('projenrc.json default project', () => {
@@ -15,7 +15,7 @@ test('projenrc.json default project', () => {
 
 test('projenrc.json with typed options', () => {
   // GIVEN
-  const project = new TestProject(renderProjenNewOptions('projen.TypeScriptProject', {
+  const project = new TestProject(renderProjenNewOptions('projen.typescript.TypeScriptProject', {
     staleOptions: {
       issues: {
         daysBeforeStale: 100, // number, nested option
