+++ test/vscode-launch-config.test.ts
@@ -1,6 +1,6 @@
 import { FileBase } from '../src/file';
 import { synthSnapshot, TestProject } from '../src/util/synth';
-import { InternalConsoleOptions } from '../src/vscode/launch-config';
+import { InternalConsoleOptions } from '../src/vscode';
 
 const VSCODE_DEBUGGER_FILE = '.vscode/launch.json';
